[Review] Code Review of #1363 (#1366)
